requirements.txt
pydub
Pillow
